DeleteConfirmationPopUp

+++ src/main/java/main/SearchView.java
@@ -90,10 +90,14 @@ public class SearchView {
     }
 
     private static void handleContextMenu(Song song, Node target) {
-        deleteSong.setOnAction(e -> {
-            HomeController.deleteSong(song);
-            songOptions.hide();
+        MenuItem deleteSongMenuItem = new MenuItem("Delete");
+        deleteSongMenuItem.setOnAction(e -> {
+            boolean confirmDelete = DeleteConfirmationPopup.show(song.getTitle());
+            if (confirmDelete) {
+                HomeController.deleteSong(song);
+            }
         });
+        ContextMenu songOptions = new ContextMenu(deleteSongMenuItem);
         songOptions.show(target, Side.TOP, -20, 5);
     }
 }
